Move TransitivityChainGraph
