+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/BenchmarkTests.java
@@ -22,7 +22,7 @@ import ai.grakn.GraknGraph;
 import ai.grakn.concept.Concept;
 import ai.grakn.graphs.DiagonalGraph;
 import ai.grakn.graphs.MatrixGraphII;
-import ai.grakn.graphs.TransitivityChainGraph;
+import ai.grakn.test.graphs.TransitivityChainGraph;
 import ai.grakn.test.graphs.TransitivityMatrixGraph;
 import ai.grakn.graql.MatchQuery;
 import ai.grakn.graql.QueryBuilder;
